feat(sequencer): embed parsing in builder
